Improve PROBE() output, enviroment var for probing fail()s
This makes PROBE() able to give more output for more types, e.g. if
given a C string it can tell you that and output it.  It will include
the pointer itself, as well as tell you what type of thing it saw
(including ENDs, freed series, trash, etc.)  For instance:

    PROBE(Context Varlist, 0x7f0fc812f8c8): tick 313419
        ../src/core/c-error.c:261

It also adds an environment variable R3_PROBE_FAILURES, which will
tell Fail_Core() to PROBE() anything it gets passed.  This is in
particular helpful in debugging boot, before any

Also moves BREAK_ON_TICK() into %sys-core.h by the include of the
debug_break() header, instead of %sys-do.h so it can be used earlier
if necessary.  Adds BREAK_NOW() as well, so you don't have to say
BREAK_ON_TICK(TG_Tick).

For good measure, aliases C-DEBUG-BREAK to C-BREAK-DEBUG, because it's
easy to get it wrong and that makes a frustrating debug session (which
probably necessitated using it in the first place) even more annoying.