Improve granularity of debug checks, PROBE() uses printf()
The previous way the debug build worked was an "all or nothing"
switch...you either had a giant build with all debug features and
asserts, or you had a release build with none of them.  With emscripten
this creates problems because the build with all the checks in it is
quite large, to the point of possibly causing the build to choke.

This breaks out the debug switches into DEBUG_XXX_YYY options.  The
default for a debug build is to enable them all and a release build
to disable them all, however they may be selectively turned back on.

Further steps will be needed, since the NDEBUG switch drives whether
assert() works at all, and some of the DEBUG_XXX_YYY features need
the assert macro to be working to be useful.  However, this focuses on
some of the data used by panic() and other routines for starters.  In
the future it needs to be addressed how to break down asserts by
category so that turning on the assert() macro doesn't turn on too
much code for smaller builds to handle.

Also decouples PROBE() from the Debug_Fmt() routine, which went through
Rebol's I/O system.  Instead it makes its own mold buffer, converts
it to UTF-8, and runs it through printf().  This simplifies the code
that PROBE() triggers to run, and means it can be used to debug the
writing-to-stdout code itself.