Fix infinite recursion in error PROBE(), unterminated mold buffer
This contains some fixes from trying to enable just DEBUG_HAS_PROBE
in the emscripten build--with no other debug flags.  It also notices
when there is a stack overflow error in MOLDing, and does not attempt
to honor the mold request coming from R3_PROBE_FAILURES in that
particular case.

There was a situation where Scan_Quote_Push_Mold(), a service routine
in the scanner, would not terminate the mold array in the case of
returning an error.  Aggressively molding every error raised during the
test suite ran across this problem, so this resolves that.