Disable stack overflow detection in emscripten, fix rebEval()
The method that Rebol used historically in recursing C routines to
detect an overflow of the C stack would measure whether successive
variables were laid up or down in memory, in terms of their respective
addresses.  This concept is not part of the C standard, which permits
abstract implementations (such as those that allocate each stack
frame for a function call on a heap).

So far Rebol has mostly been built for conventional compilers, that
just grow C stack frames up or down.  But emscripten is weird, and so
the answer of how to do stack overflow detection on it is going to be
implementation-specific.  For the moment, this disables checking
until a more suitable way is found.

Also, there was a problem when two rebEval() instructions were used in
a row.  The data was being copied out of the instruction into the
frame's temporary cell, so that the instruction could be freed.  But
the evaluator has a need to hold two instructions in its hands at
the same time in order to do one unit of lookahead.  If both were
rebEval() instructions, then both would try to lay claim to the
temporary cell...and one would be overwritten.

This modifies the protocol so that the instruction is held onto for
one evaluation cycle, where f->value points into the singular array
that represents the instruction at its single value cell.  A DO_FLAG
is set to tell the frame that f->value is such a reference, so that
if a GC runs then the singular array can be derived from the value
and marked for preservation.  Then, on the next fetch, if that DO_FLAG
is set then the data can be copied into the frame cell, and if there
is a new rebEval instruction it can now occupy the f->value slot.

To do this requires the lookahead process to be able to understand
that it can't hold onto f->value across a Fetch_Next_In_Frame(), but
it must take a possibly changed (or possibly the same) pointer back,
which will be to the same data that was in f->value but *maybe* at
a different address.

Adds a rebMold() and rebInteger() API for the test program.