Allow right of AND/OR to be void, enfix XOR, remove NOR/NAND
AND and OR do not allow voids on their left side, which they use to
determine the type of the result (LOGIC! means the result will be
a LOGIC!, otherwise it will use BLANK! to return a false result...or
chain the last evaluated truthy value).  However, due to their flexible
role as the "logic-driven-analogue-to-ELSE-and-ALSO", it is desirable
that the right hand side not be forced to a value.  For instance:

   find haystack needle or (print "Didn't find the needle")

That's an alternative way of writing:

   any [find haystack needle | print "Didn't find the needle"]

Which takes advantage of the natural infix separator ability of OR for
readability, and helps emphasize the driving operation "up front"
instead of having the user first confront an ANY.

This makes voids act as blanks for purposes of the result, when the
right hand GROUP! evaluates to one.

Also, while technically possible to implement as enfix operators in
the same family, NAND and NOR are not particularly good fits.  The
reason is that one expects some indication of the negation on the left
of the expression, like `if neither (...) nor (...)`.  Just seeing
`if (...) nor (...)` doesn't seem right.  This commit removes those
until further notice.

Fixes an issue in EITHER-TEST which was treating voids as LOGIC! false