Fix RECYCLE/TORTURE, GC queue of singular array
An assertion would occur if a GC happened while the frame was holding
a rebEval() instruction from a va_list.

Also re-enables the /TORTURE option of Recycle, which recycles on
every evaulator step.  Previously it set the GC_Disabled flag as part
of its operation, but the GC was modified (at some point) to interpret
this as the GC not being enabled at all...making the intent of the
switch useless.