Move OPEN? and LENGTH(-OF) into REFLECT method of TLS, HTTP
The scheme by which **length of x** can be written is that OF quotes
its left argument and uses it as a "reflector".  This is how R3-Alpha
actually worked to implement WORDS-OF and some other things, by
having them call REFLECT with 'WORDS and dispatch that to objects.
It was not how it did other things, for instance LENGTH or OPEN?.

While the internal routines were changed to take LENGTH and OPEN? away
from being their own methods, this never made it out to the TLS and
HTTP scheme.  Hence they would error with LENGTH OF PORT or OPEN? PORT.

This changes them, which fixes things like `read https://github.com`.
The bigger question of what the difference between a "reflector" and
an "action" is--and how to improve the discoverability--remains an
outstanding issue in port/action design.