Flexible panic locations w/"Touch" Series, limit enablement
The ability to Panic on a Series has long had the ability to identify
the stack trace of where that series was allocated, if one is running
under Address Sanitizer or Valgrind.  The simple trick is to perform
a malloc and a free on an unused integer field, and then at the moment
one wishes to query the origin of a series to dereference this "free"
memory to get the desired trace.

It's a trick that works well.  But it can also be interesting to note
other moments in a series lifetime for strategic debugging.  This
commit provides the primitive Touch_Series(), which can be used to
update the freed memory guard at any moment on a series, so that if
it is panic'd on later it will identify that location.

Also, this commit makes it so that the feature is only enabled by
default if one is running Address Sanitizer.  This keeps debug builds
that are unlikely to use the feature from bothering with the malloc
and the free on every series creation/destruction.