Change Visual Studio EXE target to "r3-exe" instead of "main"
A recent change put %host-main.c into its own "project" so that most
of Rebol could be built without it.  This was then called "main".
However, that collided with the name that the Visual Studio project
used to refer to the linker-only project that you would build to
ask for the executable.

This changes the name of that linker step project to "r3-exe"