Move TRIM to userspace, fix bug in MAKE BITSET! for BLOCK!s
Functionality for TRIM was written in C code.  But as tasks go it's
the sort of thing that if PARSE isn't able to do efficiently, then
something is wrong.  The code would have to be ported to UTF-8
everywhere's string details, which isn't as good an investment of
time as moving it to userspace and focusing on making PARSE better.

This is a first cut at an implementation which can certainly be
improved.  Additionally, it is observed that the TRIM on OBJECT!
functionality was not thought through w.r.t. to functions referring
to the binding of the original object and not the trimmed one.  The
behavior is carried forward as-is, though the feature is questionable.

Also fixes a bug in MAKE BITSET! on a BLOCK! with INTEGER! elements.