Add char* aliases in unions for series data
Some debuggers (e.g. GDB) don't offer a view of character data as UTF8
unless the data type is specifically `char*`.  This conflicts with
the desire of Rebol to manage what could be arbitrary byte data with
REBYTE*.

Because it's not that hard to do, this makes the series store char*
pointers which are then later cast to REBYTE*.  This lets one drill
down to the series contents in the debugger and see the string--*if*
the string happens to contain UTF-8 data.  This is helpful for doing
debugging of UTF-8 Everywhere.