Move build spec into extension directory
This makes each extension more self-contained.

As the extension block is auto generated by discovering make-spec.r
files under %src/extensions, the sequence might not satisify the
extension dependency, e.g. Locale depends on process for GET-ENV, but it
might be discovered before process, if the directory is iterated
alphabatically. To solve this problem, a REQUIRES property is introduced
to the extension class, and it could be a block of words (the name of
the extension). When the REQUIRES property is set, the sequence in which
the extensions are loaded will be calculated properly by make.r.