Fill out RXT_xxx type table to cover all current types
The only part of R3-Alpha's RL_API types and code that remains is the
decision to have a separate datatype enumeration, RXT_XXX values,
from the internal REB_XXX values.  So RXT_BLOCK would be a different
number from REB_BLOCK.

This was based on the idea of keeping binary compatibility for the
API across a rearrangement of order of the internal types--or even
permitting the emulation of types that were removed, to keep old
extension or API client code working.

However, the types did not have full coverage of Rebol types--at any
time.  (e.g. there was no RXT_MONEY, or RXT_ERROR.)  This updates the
types to have an entry for every current Rebol internal type, though
it deliberately shuffles the ordering--in particular to discourage the
use of type-range tests on the values (e.g. testing for something being
an ANY-ARRAY with >= RXT_BLOCK and <= RXT_PATH or similar ideas), as
that would become broken if new types were added.