First cut demo of intermixing UTF-8 with REBVAL* in rebDo()
This is a *very rough* proof of concept, showing for the first time
a functional form of intermixing REBVAL* API pointers and UTF-8
text to a parameterized rebDo.  This pattern permits flexible
combinations of text runs with spliced content, e.g.

   rebDo("print [", value1, "+ (2 *", value2, ")]", END);

One central trick is the usage of platform-independent byte ordering
macros, which are able to make the pattern of valid UTF-8 strings
distinct from the pattern of Rebol values (also distinct from
Rebol series, END markers, freed cells, etc).

There are numerous issues to resolve: one is how binding will be
handled (both for the string runs and for the spliced values).  But
also the semantics of evaluation, regarding whether WORD! or FUNCTION!
values will need to be live by default and QUOTEd or rebQuote()d...
or if they will be inert and brought to life via EVAL/rebEval().

However, this shows the general direction of what is possible and
tests the methods for how various problems might be attacked.