Adjust ONCE-BAR, allow OR and AND to accept BLOCK! on right
The mostly-forgotten "once-bar" experiment came from the first days of
variadics.  The idea was to create a form of expression barrier which
would only allow one complete expression to run, before expecting
another barrier.

    >> 1 + 2 || 3 + 4 || 5 + 6
    == 11

    >> 1 + 2 || 3 + 4 42 || 5 + 6
    ** Error: || expected single expression, found residual of 42

There are a few things preventing a "proper" implementation of the
idea (inability to measure or affect left expressions, acting as
an expression barrier is not something userspace can quite do...)
However, it had atrophied to the point of asserting.

This patches it so that it works as well as it ever did instead of
crashing.

Also includes a change to make conditional AND and OR accept a BLOCK!
on the right as well as a GROUP!.  The change makes it so that use of
a GROUP! signals that a pure LOGIC! result is desired, while a BLOCK!
suggests wanting a result that is more akin to what an ANY or ALL
would do...passing through the last evaluated value if a match, or
BLANK! on a failure.