Disallow voids when GROUP! used with AND/OR, update comments
For the sake of being conservative, this disallows the right hand
side of AND/OR evaluating to void when used with a GROUP! on the
right.  The right hand side must be allowed to return void when it's
a BLOCK!, but it's less convincing that the GROUP!/logic case needs
the ability.

Both cases require the left hand side to be non-void, as allowing an
OR with a BLOCK! on its right to have a void left hand side would
make it a poor "conditional" operator.