Equivalent to R3-Alpha console binding for rebDo()
This makes the behavior of rebDo() roughly equivalent to if you'd
typed the code into the console.  All code (including REBVAL* that are
spliced) is bound into the user context, with a fallback that new
bindings are connected to the lib context.

(Not a general solution, but just to address usage for the moment in
usage of the JavaScript build.)