Get rid of `f->kind` debug information
`f->kind` was a debug-only cache of the type of the current `f->value`
in the frame.  It became noisy to try and maintain this cache as more
places could potentially set f->value.  As debug caches go, it hadn't
turned out to be all that terribly useful.

Includes a few comment updates.