Use libRebol calls to build argv BLOCK!, vs internal API
Originally, %host-main.c was not linked against the "core" internals,
rather the "RL_API".  The API was rather limited, and so it was hard
to change the console.  It was instead linked against the %sys-core.h,
so it had full access the the internals of Rebol.

With time, the console was factored out to be mostly usermode code.
The new libRebol replacement of the old RL_API is now emerging, and
should eventually be complete enough to be the only API the remaining
C code behind the console needs.

This moves another step closer to that point by using libRebol instead
of the internal API to build the BLOCK! of STRING! version of the
argv[] passed into main().