END ignores values to end of block/group, simplify invisibles
This adds a new usermode function which consumes items in a way
that makes it appear evaluation of a BLOCK!/GROUP! has ended:

    >> do [1 + 2 end 10 + 20 | 100 + 200]
    == 3

It does this by being a hard-quoted variadic, that just takes its
arguments without looking at them...and then it is also an "invisible".

Testing this exposed a difference between:

    >> do [1 +]
    ** + is missing its value2 argument

    >> do [1 + end]
    ** + does not accept void as its value2 argument

That was because the internals of Do_Core() would convert ENDs into
voids too early for argument fulfillment to tell the difference.  This
commit changes it so that the conversion to a void is done later by
those callers of the evaluator that need it.