Error instead of assert on "invalid port actions"
R3-Alpha had a mechanism for turning BLOCK!s into PORT!s, which was
applied on a certain list of actions.  This was done based on the
numeric ordering of those actions.  A change in Ren-C scrambled this
order by no longer having "action numbers" but using symbol pointers.

Hence any actions on blocks that were not understood would be passed
to the PORT! dispatch code, causing assertions in things like addition
or WORDS-OF.  This turns that assert into an error, but the underlying
issue still needs a thought-out answer in the port model.

Also cleans up some trailing whitespace.