Make #noreturn annotation for libRebol API functions
Changes the macros for invoking rebFail (and other functions that
should not return) to be inline functions which wrap the call to
the interface table.  That's because it's not possible for function
pointers in a struct to carry the "ATTRIBUTE_NORETURN" annotation.

Note: Doesn't currently support multi-line parameter specs.

Also turns some DO MAKE ERROR! instances into FAILs.