Discern API values from non-API values via NODE_FLAG_ROOT
Previously there was no way to discern an API `REBVAL*` from one that
lived on the C stack (e.g. DECLARE_LOCAL) or inside an array cell.
This combines VALUE_FLAG_CELL with VALUE_FLAG_ROOT, assuming that all
root cells are pairings governed by the API rules.

This solves a previous problem of not knowing when it would be safe
to use non-stack-cells as a destination for evaluations (because if
evaluations were performed into an array, the memory might be moved
by array growth or shrinking via arbitrary code).  NODE_FLAG_ROOT
can be another valid condition, along with CELL_FLAG_STACK.

Also renames several flags that apply to cells with CELL_XXX instead
of VALUE_XXX.