rebDo() => rebRun(), introduce rebElide()
`rebDo("%foo.reb", END)` would be expected to execute a script, and
act as `do (%foo.reb)`.  So this is separated from the idea of just
"running" a stream of code in the evaluator, which is named rebRun().

rebDoValue() is kept for the moment, in order to avoid confusion by
renaming it to rebDo() and making it variadic...letting this settle
first.  Though even still, in the interest of cutting down on too
many API entry points, It may be that users are asked to use
`rebRun("do", ...)` instead of `rebDo(...)`.

Also adds rebElide(), which is similar to rebRun(), but returns void.
This saves on a handle allocation and the associated worries of the
returned handle's lifetime.

While going through and updating the comments mentioning rebDo(),
also corrects some outdated information.  Excessive commenting in
QUOTE is cut down as well.