rebRescue() for abstracting exception handling
It is desirable when building as JavaScript (or C++) to be able to
use the in-language exception-handling mechanism, as opposed to setjmp
and longjmp--which are relatively inefficient, and error prone.
(Emscripten must emulate longjmp painstakingly with JavaScript throws
and C++ has "zero cost exceptions").

This makes a first try at abstracting the exception handling via an
API function that takes a pointer to a C function (which may be
a JavaScript function if compiled as such)--and protects its execution.
Introduces some usages (e.g. in TRAP and REMOVE-EACH).

This is distinct from what rebTrap() would be (running Rebol code).
So the method is called rebRescue(), to name it after a parallel
function in the Ruby codebase.

Includes crash patch in FOR-SKIP when the supplied word is void.