Clarify VECTOR!'s use of series MISC() field
While the vector code isn't all that promising, it has been kept around
for its usage in the FFI and its potential to show what would need to
be done for a proper user-defined-type extension.  Such an extension
would not have access to the internal bits of series nodes to pack
this information in.

This goes ahead and eliminates the "vector type" enum, and uses
bitfields to make what's going on a bit clearer.