Add rebRescueWith(), make behavior more parallel to TRAP
Initial introduction of rebRescue() had the C function which was
"rescued" (in the same way Rebol code could be "TRAPped") with no
return value.  Rather than have it be void, this changes it to return
a REBVAL*.  Clients can choose whether to have this be NULL or not.

It also adds a rebRescueWith(), which parallels TRAP/WITH by giving a
C function handler to call.  Just as with TRAP, the presence of a
"with" handler suppresses the logic which turns normally returned
errors to voids and voids to blanks.

This produces a slightly more convoluted contract,for the sake of
convenience--though it's no worse than TRAP.  Clients wishing to opt
out of the flexibility can just return NULL from their hooks.