Use SER_AT_RAW() instead of BIN_HEAD() to pass filename
In the soon-to-be-deprecated model of exchanging data with "devices",
raw pointers to filename data are passed to the code that deals
directly with the OS.  The code either expects byte-sized series or
wide-string series.

The RENAME operation was using BIN_HEAD() to extract the filename,
which asserts that series are byte sized.  As a near-term patch this
uses SER_DATA_RAW(), and asserts a size of 1 or 2.

Code is completely different after UTF-8 Everywhere--where the unit of
exchange with the "device" is the FILE! REBVAL itself, and information
extracted from it at the last minute before the OS call.