PANIC (vs. FAIL) when natives return wrong type in debug build
There was a check of the native return types (which also double-checks
any other dispatcher type).  In theory this would help enforce the
return type contract of natives, which the dispatchers dont' check
(since those are the native functions themselves).

However, it was merely raising a normal error, instead of asserting or
otherwise raising a "this is a bug" report.  This commit does a small
bit of reorganization and switches it to a panic.  It also makes this
particular debug switch independent: DEBUG_NATIVE_RETURNS