Kill DON'T experiment, turn into DO_FLAG_NO_PATH_GROUPS
The somewhat-cutely named "DON'T" operation was a variant of DO which
could skip across code and locate the end point of a DO/NEXT operation,
but without side-effects.  While it was a given that it could not
*always* do this, some operations were thought to be scannable.  But it
really could only be used in such narrow circumstaces (e.g. a string of
literals and GET-WORD!s) that it wasn't useful.

This removes the operation and much of its supporting code, except for
what was used to prevent the evaluation of a GROUP! in a path during
something like a GET or a SET.  This prevents:

    >> path: 'obj/(print "side effect!" 'key)

    >> get path ;-- not obvious this would have side-effects
    side effect!
    == ...

Instead users are asked to use EVAL or REDUCE to accomplish GET or
SET when the PATH! will have side effects, and this raises an error.