Some %host-main.c streamlining
The method for making a Windows program be either a console app or a
GUI app depending was tied into the main() function a bit too tightly.
This breaks that logic out into its own function so it's easier to
read the flow of basic initialization.

Also, since /View would be an extension under the new model, the
relevant initialization can be done in that extension.  The one thing
it might need is the App_Instance, which is still captured.