Make HALT use THROW mechanic (vs. FAIL's longjmp)
The mechanical distinction between raising an ERROR! (e.g. FAIL) vs.
"throwing" is that throws are cooperative, returning values up the
call stack by way of the output slot of functions...while a FAIL is
done with a longjmp.

From a user's point of view, the distinction is that you intercept a
FAIL with a TRAP, and a THROW with a CATCH.  Throws are not considered
to be errors, because you can say `trap [... if x [throw 10]]` and
the TRAP will not catch it.  Also, a CATCH must know the name of a
named throw in order to catch it.

This would make it seem that HALT is more like a THROW or QUIT or a
RETURN...because it is intended to jump up to the top level of the
console...bypassing any TRAPs on the stack.  But for mechanical reasons
R3-Alpha used the longjmp mechanism to process it.  This is because
there was no connecting path from the locations that polled for the
HALT to an output slot to write it to.

As a result, HALT became the one kind of special failure that would not
be trapped...by any traps but the topmost level.  This gave rise to
PUSH_UNHALTABLE_TRAP() and PUSH_TRAP() as distinct macros, and also
introduced concern into those trapping failures who were not involved
in Rebol's control flow (e.g. piping a D_OUT) that a halt might jump
across them if they didn't use PUSH_UNHALTABLE_TRAP().

This changes HALT to use the throwlike mechanism, which required making
the wait routine on ports being changed to be able to return a thrown
value in the case of interruption.  The result simplifies matters,
though it does interfere with a trick used in Ren Garden for running
on more than one thread (which would have to be revisited along with
many other API changes anyway).