rebLastError() -> rebTrap(), rebRescue(), host exceptions
The first try at the libRebol API used a GetLastError()-style approach,
where receiving a NULL value handle from a rebDo() API call would
signal that an error had occurred.  This was fairly weak in terms of
how it could be made to integrate with the exception model of a host
that might support try/catch (like C++ or Java)...and it precluded
the use of functions that mixed evaluations with unboxings...where
no obvious place to indicate a failure would exist.

This goes ahead and rips it out.  It also gets rid of some kind of
ad-hoc conventions that were used to indicate a HALT had been received
by signaling a BAR! as the "error" or an INTEGER! as the "error" for
a QUIT request...instead letting these unhandled thrown-type-things
be processed by interpreting the error which showed them as unhandled.
That leverages the change made to use the throwing mechanic for HALT
instead of a longjmp, and avoids exposing the thrown-bit mechanics
to the API.

It also kills off what appeared to be some dead stdio code, for a
parallel stdio device implementation.  (Apparently some kind of
"simple IO" pre-Rebol boot, which we'll assume can be taken care of
by printf() in debug builds.)