Some preservation of new-line flag on REDUCE/COMPOSE
This does a simple level of preservation of newlines when a REDUCE
or a COMPOSE is performed.  If a GROUP! is evaluated in COMPOSE, then
if it has a newline set that will be added to the first spliced
element at that position.  If the leading item in an evaluation for
a REDUCE has a newline set then it will be added to the evaluative
result at that position.