Reallocate series flags by separating out array-only flags
Ren-C increased the size of series nodes from 6 platform pointers to
8, and to make up for it changed it so more things could fit into
series nodes in their entirety.  The way this was done also opened
up more opportunities for series flags.  However, even though there
were many more flags available they were running out...in particular
the SERIES_FLAG_XXX group which could be passed to Make_Series() to
control series creation.

One limitation of how many flags could be used was that the rightmost
16 bits of the header flags (in the first platform-pointer-size slot)
were being used by UTF-8 data for ANY-WORD! strings to encode an
optional symbol number.  This slot was left open for other series
types, such as arrays to use...though no usage for a 16-bit number
had come up.  Since having more flags is more valuable, this changes
it so that those 16-bits are used to store array-specific flags.

The upshot of this is that if you have a pointer to something you
know is a series but you don't know if it's an array, you can't test
for an array-only bit without also testing for SERIES_FLAG_ARRAY.
This is currently a very uncommon scenario, however.

As a side-effect of this process, it also frees up one generic node
flag that had been somewhat abused by calling it NODE_FLAG_SPECIAL
and then using it for different purposes in series vs. cells.  A
more appropriate purpose would be a test that would be applied when
not knowing if something was a series or a cell yet being able to
make a decision based on the finding.  It is now reserved for future
applications of that kind.