Fix semiquoting left of enfix, error leak on shutdown
The unevaluated bit was not being preserved when a value was not
quoted and moved from the output slot into a frame argument to fulfill
a left enfix argument.  This corrects that, and uses it to check to
make sure the left hand side of OR/AND/XOR is properly conditional
and not accidentally being used as `[1] or [2]` or something that
misunderstands the left hand vs. right hand distinction.

Also fixes a shutdown leak, and an infinitely recursive panic problem
when a panic is performed on an END marker.