Move backtrace into usermode
This takes the BACKTRACE routine and rewrites it in Rebol.  It was
converted from the C mostly line-by-line just to get it started.

It's really just a starting point.  But getting it to work at all
required making it so that it was possible to climb the frame list,
so a routine called PARENT-OF was exposed for this purpose.  It was
also necessary to be able to speak of a FRAME! that was on the stack
but whose variables weren't accessible yet (because only some arguments
were fulfilled), so tentative support for that was added.

Moving this to usermode meant it was not necessary to have some
code for popping the data stack in reverse, which was only used by
the native backtrace.  This removes the C code for the sake of
simplicity.