Allow literal non-blocks as left of infix OR, AND, XOR
The current rule for conditions in conditionals is just that they not
be void and not be literal blocks.  This avoids cases like turning a
WHILE into an IF and getting `if [a < b] [...]` to deceptively run
with a "true" condition.

But plain literals like 1 or <foo> are permitted.  While this might
seem useless enough to warrant telling you not to do it, the odds of
doing it accidentally are much less than with BLOCK!.  And in REDUCE
or COMPOSE scenarios, it may be desirable to let literal conditions
through.

So this eases off a restriction on the left hand side of OR/AND/XOR to
allow literals besides BLOCK!s, bringing it in line with the rule
for conditions.