Work around "multiple inheritance" use in HELP
R3-Alpha permitted MAKE to take an OBJECT! as both its arguments, and
it would do a kind of "multiple inheritance" on the result.  This is
semantically very dodgy, because it would mean pulling two unrelated
objects together in such a way that they could incidentally stomp on
each others members in a hard to decipher way.  But mechanically
speaking, it prohibits "derived binding" in its current form.

HELP was using the feature as a way of avoiding repetition in listing
items out of user and lib contexts.  It would unify the contexts into
one where user overwrote lib.  This feature too is somewhat haphazard,
but this commit preserves it for now by doing APPENDs.

Also removes an old experimental feature by which the backtrace number
for a function was the interpretation of something like SOURCE 3.  That
would get you the source of the function at backtrace index 3.  The
method is changing, and if the debugger adds this ability to the
common SOURCE function that should be done by another method.