MAKE PREP without need for MAKE CLEAN for mezzanine
Because C only lets you initialize arrays with compile-time constants,
some macros were used to talk about the embedded natives and code.
These could get out of sync and commonly lead to boot errors when
a change was made to the mezzanine.

This adds extern const values for those macros to be linked to instead
of hardcoded into the callsites, so that recompiling the file with
the C literal data and relinking is enough to avoid the boot error.