Use rebRescue() to implement SCAN_FLAG_RELAX
A current goal is to re-stylize all the PUSH_TRAP/DROP_TRAP pairings
into using functions like rebRescue().  This way, it would not matter
if the system was using setjmp/longjmp or C++ try/catch or JavaScript
try/catch under the hood to implement failures.  It would permit
using more efficient models.

One tricky case was how PUSH_TRAP was used conditionally based on a
flag to the scanner.  What made it tricky was that it expected to be
able to react to the failure while still having access to partially
constructed arrays on the data stack.  (Typically a FAIL will roll back
the data stack to the state it was in at the time of the PUSH_TRAP.)

It would be slow to make every step of the scanning operation be
guarded by a rebRescue() in case it failed, to catch errors that are
happening on a relatively uncommon basis.  So the trick this uses is
to only use a rebRescue() on array scans where the SCAN_FLAG_RELAX
is set...catch the scanning failure, and then scan it *again* but only
up to a point that had prior been a success.

Patches a build error from the previous commit introducing a stray
space in a comment-block definition for REBNATIVE(compile).