Add a user script hook to WAIT
Now user can insert code into system/ports/pump if it needs WAIT to poll
something (e.g. calling select or uv_run, etc.) and pump events into the
system. This makes it possible to integrate other event loops.

As a note, because the hook (a block) is executed in WAIT at each
iteration, it should finish as soon as possible, or it will slow down
I/O of the whole program.