Export rebMalloc(), don't use PUSH_TRAP in inflate/deflate
In the efforts to try and abstract out the failure mechanism (so that
whatever throw/catch mechanism could be used, besides just setjmp and
longjmp), each usage of PUSH_TRAP is being reviewed.

Because zlib has internal state which needs freeing in the event of
a failure (e.g. running out of memory when resizing the target series),
there had been a PUSH_TRAP in the code doing inflating.  If an error
occurred, the code that did the trap would call inflateEnd() before
re-triggering the error.

Rather than blindly convert this into a rebRescue() call, this evolves
code that was created for LodePNG, which tells zlib to use a special
allocator rebMalloc(), which uses series as the backing store for
the data.  Originally this was created for the purpose of not needing
to do a reallocation when memory was eventually targeting a series
anyway.  But in this case, the advantage of backing zlib's internal
state with a series is that this manual series will be automatically
freed when a failure occurs.

This reviews and improves the rebMalloc() interface, so it is used
more cleanly inside of zlib.  It also changes the rebXXXAlloc() based
routines to use rebMalloc() instead of OS_ALLOC to acquire memory.