rebMalloc(0) does not return NULL, poison less underrun
The idea that the result of a rebMalloc() can be passed to functions
that manage generic properties for asking the handle to change its
lifetime management, or for instance rebRepossessed, suggests using
the standard's wiggle room about the definition of rebMalloc() to
not return NULL... so that there are fewer special cases to handle
in those other routines.

Also, since REBVAL* and these generic buffers are being brought
together in the API, this means that singular arrays will be used for
API values instead of pairings.  Hence it won't necessarily be a good
idea to poison ALIGN_SIZE worth of bytes before the memroy, as only
a platform pointer's worth of data may precede the data.  This cuts
down how much is poisoned to just one pointer's worth.