API REBVAL* managed by default, paired => singular
It was a goal for the API to be "easy to use" at the potential cost of
inefficiency.  So an API-generated REBVAL* could be allocated, used,
and left to be automatically cleaned up when the function frame it was
allocated in ended.

With API use growing in the system, however, this would invite a kind
of "leak" in code that would like to know if it was "leaking"--in the
sense of burdening the garbage collector when it could be releasing
handles manually.  Also, some top level C code (such as the console)
would run indefinitely, meaning it would need to do lifetime management
of its handles.

So as an interim measure, API handles were unmanaged.  This commit
changes it so that they are managed, however the GC will panic if it
has to free them in cases that aren't a fail().  So this doesn't change
much except it is no longer needed to rebManage() a handle if you plan
to call rebFail().

It also switches the API handles to live in singular REBSER instead of
"pairing" REBSER.  It really is not strictly necessary that REBSER
nodes be used at all for the containers of API handles (they could be
in their own pool managed elsewhere, which would also make them easier
to walk as roots without needing an independent pass).  But making them
the same size cuts down on definitions and enables greater reusability.

The reason to switch from a pairing to a singular array is in order to
locate a REBNOD* prior to the value payload, in the link field.  These
"underrun" bytes can then be examined to discern a rebMalloc() data
allocation from a REBVAL* handle...which cannot be done by looking at
the actual pointed-at data itself.