Eliminate %root.r and %task.r, allow API roots during boot
R3-Alpha had the concept of "root series" which held various series
alive, even though they had no references in user code.  This would
be series like the mold buffer, or something like a plain empty block
that could be used many places.

Practically speaking, these root series came down to being two arrays
that were hidden from the user.  These were the ROOT array and the
TASK array, with the concept that in a multithreaded program each
thread would get its own instance of the variables in TASK...while
ROOT would be global.

Ren-C introduced the idea of unmanaged series, and that a series would
have to be explicitly managed in order for the GC to pick it up.  This
means that for anything that wasn't meant to be exposed to the user
as a value (such as the mold buffer) it would be fine to just do a
Make_Series() call to create it, and then a Free_Series() call at the
end of the program.  By not putting the series into a REBVAL, it could
be accessed slightly faster (e.g. no need to say VAL_SERIES(mold_buf))
and it would reduce the risk that these system arrays would become
exposed to the user.

An additional mechanism also exists now for REBVAL* that want to have
an indefinite lifetime, which is to use Alloc_Value().  So even if one
wants a full value, it can be done without the ROOT or TASK arrays.

This simplifies matters by just not using arrays at all.  The Rebol
files %task.r and %root.r can be eliminated in favor of plain PVAR
and TVAR globals, and it means there doesn't need to be code generating
C structures backing the arrays in %make-boot.r